Fixed gene name KeyError issue
Adjusted the Uniprot fetch file to return at the very least a None value for the gene name if one does not exist. However, if the record is associated with an ORF it will return that name instead. These should mostly be issues associated with non-human or unreviewed records primarily.